refactor!: scope lsp options under elixirls key
When setting up elixir-ls, one should now scope all of the opts under an
`elixirls` key in the opts table.

Old:

```lua
local elixir = require("elixir")

elixir.setup {
  settings = {
    -- some settings
  }
}
```

New:

```lua
local elixir = require("elixir")

elixir.setup {
  elixirls = {
    settings = {
      -- some settings
    }
  }
}
```

In addition, the function to merge settings into the default settings
now lives in the elixirls module

Old:

```elixir
local elixir = require("elixir")

elixir.settings {
  -- some settings
}
```

New:

```elixir
local elixirls = require("elixir.elixirls")

elixirls.settings {
  -- some settings
}
```
